+++ src/boost_histogram/_internal/hist.py
@@ -27,12 +27,12 @@ import numpy as np
 import boost_histogram
 from boost_histogram import _core
 
+from .._utils import cast, register, set_module
 from ..axis import AxesTuple, Axis, Variable
 from ..storage import Double, Storage
 from ..typing import Accumulator, ArrayLike, CppHistogram
 from ..view import MeanView, WeightedMeanView, WeightedSumView, _to_view
 from .enum import Kind
-from .utils import cast, register, set_module
 
 if TYPE_CHECKING:
     from builtins import ellipsis
