+++ src/boost_histogram/_utils.py
